Revert "Добавлены классы Book и RootVegetable. Добавлены классы генерирующие экземпляр Book и RootVegetable с рандомными значениями. Переработано меню. Создан пакет для пунктов меню menu. Создан пакет models для кассов Car Book RootVegetable. Создан пакет util. Создан пакет для рандомайзеров util.randomfill"
This reverts commit b3a48ec2bd9452e9229c9f3a84f989b9db33f46d.
