+++ src/main/java/BookRandomizer.java
@@ -1,8 +1,3 @@
-package util.randomfill;
-
-import models.Book;
-import models.Car;
-
 import java.util.Random;
 
 public class BookRandomizer {
