+++ src/main/java/Main.java
@@ -1,19 +1,44 @@
-import menu.*;
+import java.util.Arrays;
+import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
 
-        Menu menu = new Menu();
-        menu.addAction(1, new ReadFromFileAction()); // Вызов меню заполнения массива из файла
-        menu.addAction(2, new FillManualAction());   // Вызов меню заполнения массива вручную
-        menu.addAction(3, new FillRandomAction());   // Вызов меню заполнения массива случайными данными
-        menu.addAction(4, new SearchAction());       // Вызов меню поиска элемента массива
-        menu.addAction(5, new PrintAction());        // Вызов меню вывода данных в консоль
-        menu.addAction(0, new ExitAction());         // Завершение работы программы
+        Scanner scanner = new Scanner(System.in);
+        boolean step = true;
 
-        while (true) {
-            menu.display();
-            System.out.println();
+        //предложение выбора пользаветюлю
+        String[] masString = new String[5];
+        masString[0] = "1 Загрузка из файла";
+        masString[1] = "2 Ввод данных в ручную";
+        masString[2] = "3 Выбор за вас";
+        masString[3] = "4 Поиск значения";
+        masString[4] = "5 Завершение програмы";
+
+        //Сам цикл с ветвлением и ввод пользователя
+        while (step) {
+            Arrays.stream(masString).forEach(System.out::println);
+            String choise = scanner.next();//взаимодействие с пользователем
+            switch (choise) {
+                case "1":
+                    System.out.println("заполнение исходного массив");
+                    break;
+                case "2":
+                    System.out.println("Временно не доступно 2");
+                    break;
+                case "3":
+                    System.out.println("Временно не доступно 3");
+                    break;
+                case "4":
+                    System.out.println("Временно не доступно 4");
+                    break;
+                case "5":
+                    System.out.println("Всего доброго");
+                    step = false;
+                    break;
+                default:
+                    System.out.println("Выбор некорректный");
+            }
         }
     }
 }
