--- src/main/java/menu/ExitAction.java
@@ -1,9 +1,0 @@
-package menu;
-
-public class ExitAction implements MenuAction {
-    @Override
-    public void execute() {
-        System.out.println("Exiting program...");
-        System.exit(0);
-    }
-}
