--- src/main/java/menu/FillManualAction.java
@@ -1,8 +1,0 @@
-package menu;
-
-public class FillManualAction implements MenuAction{
-    @Override
-    public void execute(){
-        System.out.println("Меню заполнения данных вручную");
-    }
-}
