--- src/main/java/menu/MenuAction.java
@@ -1,5 +1,0 @@
-package menu;
-
-public interface MenuAction {
-    void execute();
-}
