--- src/main/java/menu/PrintAction.java
@@ -1,10 +1,0 @@
-package menu;
-
-import models.*;
-
-public class PrintAction implements MenuAction {
-    @Override
-    public void execute() {
-
-    }
-}
