--- src/main/java/menu/ReadFromFileAction.java
@@ -1,8 +1,0 @@
-package menu;
-
-public class ReadFromFileAction implements MenuAction {
-    @Override
-    public void execute(){
-        System.out.println("Меню чтения из файла");
-    }
-}
