--- src/main/java/menu/SearchAction.java
@@ -1,8 +1,0 @@
-package menu;
-
-public class SearchAction implements MenuAction{
-    @Override
-    public void execute(){
-        System.out.println("Меню поиска элемента");
-    }
-}
